Used data in LearningPathDetails, Course and Lesson's components
